+++ src/tada-commands.ads
@@ -1,6 +1,8 @@
-with Tada.CL_Arguments;
+with Tackle.Opts;
 
 package Tada.Commands is
+   use Tackle;
+
    Parse_Error : exception;
    Execute_Error : exception;
 
@@ -8,7 +10,6 @@ package Tada.Commands is
                          Cache,
                          Config,
                          Clean,
-                         Help,
                          Init,
                          Install,
                          Run,
@@ -18,7 +19,7 @@ package Tada.Commands is
    type Profile_Kind is (Debug,
                          Release);
 
-   type Command (Kind : Command_Kind := Help) is record
+   type Command (Kind : Command_Kind) is record
       case Kind is
          when Build =>
             Build_Profile : Profile_Kind;
@@ -29,16 +30,16 @@ package Tada.Commands is
             Package_Type : Package_Kind;
          when Run =>
             Run_Profile : Profile_Kind;
-            Args : CL_Arguments.Argument_List.Vector;
+            Args : Opts.Argument_List;
          when Test =>
             Test_Profile : Profile_Kind;
 
-         when Clean | Config | Help | Install | Version =>
+         when Clean | Config | Install | Version =>
             null;
       end case;
    end record;
 
-   function Parse (Arguments : CL_Arguments.Argument_List.Vector) return Command;
+   function Parse (Result : Opts.Result) return Command;
 
    procedure Execute (Cmd : Command);
 end Tada.Commands;