+++ tests/tada_commands_tests.ads
@@ -3,15 +3,9 @@ with Testy.Tests;
 package Tada_Commands_Tests is
    use Testy.Tests;
 
-   procedure Test_Parse_Empty_Args (T : in out Test_Context);
-
-   procedure Test_Parse_Help (T : in out Test_Context);
-
    procedure Test_Parse_Clean (T : in out Test_Context);
 
    procedure Test_Parse_Version (T : in out Test_Context);
-
-   procedure Test_Parse_Foo (T : in out Test_Context);
 
    procedure Test_Parse_Build (T : in out Test_Context);
 
@@ -19,15 +13,9 @@ package Tada_Commands_Tests is
 
    procedure Test_Parse_Build_Release (T : in out Test_Context);
 
-   procedure Test_Parse_Build_Profile (T : in out Test_Context);
-
    procedure Test_Parse_Build_Unknown (T : in out Test_Context);
 
-   procedure Test_Parse_Build_Unexpected (T : in out Test_Context);
-
    procedure Test_Parse_Run (T : in out Test_Context);
-
-   procedure Test_Parse_Run_No_Args (T : in out Test_Context);
 
    procedure Test_Parse_Run_Args (T : in out Test_Context);
 
@@ -39,8 +27,6 @@ package Tada_Commands_Tests is
 
    procedure Test_Parse_Init_Name_Lib (T : in out Test_Context);
 
-   procedure Test_Parse_Init_Name_Foo (T : in out Test_Context);
-
    procedure Test_Parse_Init_Name_Camel (T : in out Test_Context);
 
    procedure Test_Parse_Init_Name_Package (T : in out Test_Context);
@@ -50,6 +36,4 @@ package Tada_Commands_Tests is
    procedure Test_Parse_Cache (T : in out Test_Context);
 
    procedure Test_Parse_Cache_Force (T : in out Test_Context);
-
-   procedure Test_Parse_Cache_Foo (T : in out Test_Context);
 end Tada_Commands_Tests;